pyproject.toml
[project]
name = "computergame"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "accelerate==1.5.1",
    "aiodns==3.2.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.13",
    "aiohttp-retry==2.9.1",
    "aiosignal==1.3.2",
    "aleph-alpha-client==11.2.0",
    "annotated-types==0.7.0",
    "anthropic>=0.75.0",
    "anyio==4.8.0",
    "attrs==25.3.0",
    "audioread==3.0.1",
    "azure-common==1.1.28",
    "azure-core==1.32.0",
    "azure-identity==1.21.0",
    "azure-mgmt-compute==34.0.0",
    "azure-mgmt-core==1.5.0",
    "azure-mgmt-network==28.1.0",
    "backoff==2.2.1",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.3",
    "blinker==1.9.0",
    "borb==2.1.25",
    "boto3==1.37.11",
    "botocore==1.37.11",
    "cachetools==5.5.2",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.1",
    "clemcore==3.4.0",
    "cli-exit-tools==1.2.7",
    "click==8.1.8",
    "cloudpickle==3.1.1",
    "cohere>=5.20.1,<6.0.0",
    "contourpy==1.3.1",
    "cryptography==44.0.2",
    "cssselect==1.3.0",
    "cycler==0.12.1",
    "dashscope==1.22.2",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "deprecated==1.2.18",
    "dill==0.3.9",
    "distro==1.9.0",
    "docker==7.1.0",
    "docker-pycreds==0.4.0",
    "easyocr==1.7.2",
    "et-xmlfile==2.0.0",
    "fabric==3.2.2",
    "farama-notifications==0.0.4",
    "fastavro==1.10.0",
    "fastdtw==0.3.4",
    "filelock==3.17.0",
    "flask==3.0.3",
    "fonttools==4.56.0",
    "formulas==1.2.8",
    "frozenlist==1.5.0",
    "fsspec==2025.3.0",
    "func-timeout==4.3.5",
    "gdown==5.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.24.2",
    "google-api-python-client==2.164.0",
    "google-auth>=2.47.0",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.69.1",
    "greenlet==3.1.1",
    "groq==0.19.0",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "gymnasium>=1.0.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "huggingface-hub==0.29.3",
    "idna==3.10",
    "imagehash==4.3.2",
    "imageio==2.37.0",
    "importlib-metadata==6.11.0",
    "importlib-resources==6.5.2",
    "invoke==2.2.0",
    "ipykernel>=6.30.1",
    "isodate==0.7.2",
    "itsdangerous==2.2.0",
    "jax-jumpy==1.0.0",
    "jinja2==3.1.6",
    "jiter>=0.10.0",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "kiwisolver==1.4.8",
    "lazy-loader==0.4",
    "lib-detect-testenv==2.0.8",
    "librosa==0.11.0",
    "llvmlite==0.44.0",
    "loguru==0.7.3",
    "lxml==5.3.1",
    "markupsafe==3.0.2",
    "matplotlib==3.7.1",
    "mistralai>=1.10.0,<1.11",
    "mouseinfo==0.1.3",
    "mpmath==1.3.0",
    "msal==1.32.0",
    "msal-extensions==1.2.0",
    "msgpack==1.1.0",
    "multidict==6.1.0",
    "multiprocess==0.70.17",
    "mutagen==1.47.0",
    "networkx==3.4.2",
    "ninja==1.11.1.3",
    "nltk==3.8.1",
    "numba==0.61.0",
    "numpy==1.26.4",
    "numpy-financial==1.0.0",
    "oauth2client==4.1.3",
    "odfpy==1.4.1",
    "openai>=2.15.0,<3.0.0",
    "opencv-python==4.8.1.78",
    "opencv-python-headless==4.11.0.86",
    "openpyxl==3.1.5",
    "orjson==3.10.15",
    "packaging==24.2",
    "pandas==2.0.1",
    "paramiko==3.5.1",
    "pdfminer-six==20231228",
    "pdfplumber==0.11.5",
    "pillow==11.0.0",
    "platformdirs==4.3.6",
    "playwright==1.50.0",
    "pooch==1.8.2",
    "portalocker==2.10.1",
    "propcache==0.3.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.3",
    "psutil==5.9.8",
    "py==1.11.0",
    "pyacoustid==1.3.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pyautogui==0.9.54",
    "pycares==4.5.0",
    "pyclipper==1.3.0.post6",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydrive==1.3.1",
    "pyee==12.1.1",
    "pygame==2.6.1",
    "pygetwindow==0.0.9",
    "pyjwt==2.10.1",
    "pymsgbox==1.0.9",
    "pymupdf==1.25.3",
    "pynacl==1.5.0",
    "pynput==1.7.8",
    "pyobjc-core==11.0",
    "pyobjc-framework-applicationservices==11.0",
    "pyobjc-framework-cocoa==11.0",
    "pyobjc-framework-coretext==11.0",
    "pyobjc-framework-quartz==11.0",
    "pyparsing==3.2.1",
    "pypdf==5.3.1",
    "pypdf2==3.0.1",
    "pypdfium2==4.30.1",
    "pyperclip==1.9.0",
    "pyrect==0.2.0",
    "pyscreeze==1.0.1",
    "pysocks==1.7.1",
    "python-barcode==0.15.1",
    "python-bidi==0.6.6",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.1.2",
    "python-dotenv>=1.1.1",
    "python-liquid==1.13.0",
    "python-pptx==1.0.2",
    "pytweening==1.2.0",
    "pytz==2025.1",
    "pywavelets==1.8.0",
    "pyyaml==6.0.2",
    "qrcode==8.0",
    "rapidfuzz==3.12.2",
    "regex==2024.11.6",
    "requests>=2.32.2",
    "requests-toolbelt==1.0.0",
    "retry==0.9.2",
    "rsa==4.9",
    "rubicon-objc==0.5.0",
    "s3transfer==0.11.4",
    "safetensors==0.5.3",
    "schedula==1.5.58",
    "scikit-image==0.25.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "seaborn==0.12.2",
    "sentry-sdk==2.22.0",
    "setproctitle==1.3.5",
    "shapely==2.0.7",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soundfile==0.13.1",
    "soupsieve==2.6",
    "soxr==0.5.0.post1",
    "sympy==1.13.1",
    "tabulate>=0.9.0",
    "threadpoolctl==3.6.0",
    "tifffile==2025.2.18",
    "tiktoken==0.9.0",
    "tokenizers==0.15.2",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "tqdm>=4.66.3",
    "transformers==4.35.2",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
    "uritemplate==4.1.1",
    "urllib3==2.3.0",
    "wandb==0.19.8",
    "websocket-client==1.8.0",
    "werkzeug==3.1.3",
    "wrapt==1.17.2",
    "wrapt-timeout-decorator==1.5.1",
    "xlsxwriter==3.2.2",
    "xmltodict==0.14.2",
    "yarl==1.18.3",
    "zipp==3.21.0",
]

[tool.ruff]
line-length = 150
